[5.3] Move render movies logic to presenter
